+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
 * ChaosScorer._score() returns the existing raw score before computing a new one
+* Filter out Japanese hashtags unless Config.language is "ja"
 
 ### v0.30.3
 * Special `mediaAttachment` repair for bluesky bridged toots